requirements.txt
Django==3.0.7
django-rest-swagger==2.2.0
djangorestframework==3.11.